@@ -8,15 +8,15 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^17.3.6",
-    "@angular/common": "^17.3.6",
-    "@angular/compiler": "^17.3.6",
-    "@angular/core": "^17.3.6",
-    "@angular/forms": "^17.3.6",
-    "@angular/localize": "^17.3.6",
-    "@angular/platform-browser": "^17.3.6",
-    "@angular/platform-browser-dynamic": "^17.3.6",
-    "@angular/router": "^17.3.6",
+    "@angular/animations": "18.0.4",
+    "@angular/common": "18.0.4",
+    "@angular/compiler": "18.0.4",
+    "@angular/core": "18.0.4",
+    "@angular/forms": "18.0.4",
+    "@angular/localize": "18.0.4",
+    "@angular/platform-browser": "18.0.4",
+    "@angular/platform-browser-dynamic": "18.0.4",
+    "@angular/router": "18.0.4",
     "@mean-stream/nestx": "^0.11.4",
     "@mean-stream/ngbx": "^0.13.0",
     "@nestjs/common": "^10.3.8",
@@ -39,41 +39,42 @@
     "zone.js": "~0.14.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "17.3.5",
-    "@angular-devkit/core": "17.3.5",
-    "@angular-devkit/schematics": "17.3.5",
+    "@angular-devkit/build-angular": "18.0.5",
+    "@angular-devkit/core": "18.0.5",
+    "@angular-devkit/schematics": "18.0.5",
     "@angular-eslint/builder": "^17.3.0",
-    "@angular-eslint/eslint-plugin": "~17.3.0",
-    "@angular-eslint/eslint-plugin-template": "~17.3.0",
+    "@angular-eslint/eslint-plugin": "18.0.1",
+    "@angular-eslint/eslint-plugin-template": "18.0.1",
     "@angular-eslint/schematics": "17.3.0",
-    "@angular-eslint/template-parser": "~17.3.0",
-    "@angular/cli": "17.3.5",
-    "@angular/compiler-cli": "^17.3.6",
-    "@angular/language-service": "^17.3.6",
+    "@angular-eslint/template-parser": "18.0.1",
+    "@angular/cli": "~18.0.0",
+    "@angular/compiler-cli": "18.0.4",
+    "@angular/language-service": "18.0.4",
     "@nestjs/cli": "^10.3.2",
     "@nestjs/schematics": "^10.1.1",
     "@nestjs/testing": "^10.3.8",
-    "@nx/angular": "17.2.0",
-    "@nx/cypress": "17.2.0",
-    "@nx/eslint-plugin": "17.2.0",
-    "@nx/jest": "17.2.0",
-    "@nx/js": "17.2.0",
-    "@nx/linter": "17.2.0",
-    "@nx/nest": "17.2.0",
-    "@nx/node": "17.2.0",
-    "@nx/web": "17.2.0",
-    "@nx/webpack": "17.2.0",
-    "@nx/workspace": "17.2.0",
-    "@schematics/angular": "^17.3.6",
+    "@nx/angular": "19.3.1",
+    "@nx/cypress": "19.3.1",
+    "@nx/eslint-plugin": "19.3.1",
+    "@nx/jest": "19.3.1",
+    "@nx/js": "19.3.1",
+    "@nx/linter": "19.3.1",
+    "@nx/nest": "19.3.1",
+    "@nx/node": "19.3.1",
+    "@nx/web": "19.3.1",
+    "@nx/webpack": "19.3.1",
+    "@nx/workspace": "19.3.1",
+    "@schematics/angular": "18.0.5",
     "@types/express": "^4.17.21",
     "@types/jasmine": "~5.1.4",
     "@types/jest": "^29.5.12",
-    "@types/node": "~20.12.7",
+    "@types/node": "18.16.9",
     "@types/supertest": "^2.0.16",
-    "@typescript-eslint/eslint-plugin": "6.13.2",
-    "@typescript-eslint/parser": "^6.21.0",
-    "cypress": "^13.8.1",
-    "eslint": "8.55.0",
+    "@typescript-eslint/eslint-plugin": "7.3.0",
+    "@typescript-eslint/parser": "7.3.0",
+    "@typescript-eslint/utils": "^8.0.0-alpha.28",
+    "cypress": "13.8.0",
+    "eslint": "8.57.0",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-cypress": "2.15.2",
     "eslint-plugin-import": "2.29.1",
